Split processing out into separate file
