+++ pkg/impl.go
@@ -10,10 +10,6 @@ import (
 	"go/token"
 	"html/template"
 	"reflect"
-	"strconv"
-	"strings"
-
-	"golang.org/x/tools/imports"
 )
 
 var tmpl = template.Must(template.New("test").Parse(stub))
@@ -33,99 +29,6 @@ func hasIdentifier(ident string, node ast.Node) bool {
 	})
 
 	return found
-}
-
-func findTopTypeDecl(id string, f *ast.File) (*ast.GenDecl, *ast.TypeSpec) {
-	for _, decl := range f.Decls {
-		decl, ok := decl.(*ast.GenDecl)
-		if !ok || decl.Tok != token.TYPE {
-			continue
-		}
-		for _, spec := range decl.Specs {
-			switch spec := spec.(type) {
-			case *ast.TypeSpec:
-				if spec.Name.Name == id || getIdent(spec, 0) == id {
-					return decl, spec
-				}
-			}
-		}
-	}
-	return nil, nil
-}
-
-// findInterface returns the import path and identifier of an interface.
-// For example, given "http.ResponseWriter", findInterface returns
-// "net/http", "ResponseWriter".
-// If a fully qualified interface is given, such as "net/http.ResponseWriter",
-// it simply parses the input.
-func findInterface(iface string) (path string, id string, err error) {
-	if len(strings.Fields(iface)) != 1 {
-		return "", "", fmt.Errorf("couldn't parse interface: %s", iface)
-	}
-
-	if strings.Index(iface, ".") == -1 {
-		fs := token.NewFileSet()
-
-		pkgs, err := parser.ParseDir(fs, ".", nil, 0)
-		if err != nil {
-			return "", "", err
-		}
-
-		for _, pkg := range pkgs {
-			for _, file := range pkg.Files {
-				if hasIdentifier(iface, file) {
-					return ".", iface, nil
-				}
-			}
-		}
-	}
-
-	if slash := strings.LastIndex(iface, "/"); slash > -1 {
-		// package path provided
-		dot := strings.LastIndex(iface, ".")
-		// make sure iface does not end with "/" (e.g. reject net/http/)
-		if slash+1 == len(iface) {
-			return "", "", fmt.Errorf("interface name cannot end with a '/' character: %s", iface)
-		}
-		// make sure iface does not end with "." (e.g. reject net/http.)
-		if dot+1 == len(iface) {
-			return "", "", fmt.Errorf("interface name cannot end with a '.' character: %s", iface)
-		}
-		// make sure iface has exactly one "." after "/" (e.g. reject net/http/httputil)
-		if strings.Count(iface[slash:], ".") != 1 {
-			return "", "", fmt.Errorf("invalid interface name: %s", iface)
-		}
-		return iface[:dot], iface[dot+1:], nil
-	}
-
-	src := []byte("package hack\n" + "var i " + iface)
-	// If we couldn't determine the import path, goimports will
-	// auto fix the import path.
-	imp, err := imports.Process(".", src, nil)
-	if err != nil {
-		return "", "", fmt.Errorf("couldn't parse interface: %s", iface)
-	}
-
-	// imp should now contain an appropriate import.
-	// Parse out the import and the identifier.
-	fset := token.NewFileSet()
-	f, err := parser.ParseFile(fset, "", imp, 0)
-	if err != nil {
-		panic(err)
-	}
-	if len(f.Imports) == 0 {
-		return "", "", fmt.Errorf("unrecognized interface: %s", iface)
-	}
-	raw := f.Imports[0].Path.Value   // "io"
-	path, err = strconv.Unquote(raw) // io
-	if err != nil {
-		panic(err)
-	}
-	decl := f.Decls[1].(*ast.GenDecl)      // var i io.Reader
-	spec := decl.Specs[0].(*ast.ValueSpec) // i io.Reader
-	sel := spec.Type.(*ast.SelectorExpr)   // io.Reader
-	id = sel.Sel.Name                      // Reader
-	return path, id, nil
 }
 
 // Pkg is a parsed build.Package.
